Корректирует функцию в thumbnails.js
